model_requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bert-extractive-summarizer==0.4.2
bleach==3.1.5
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-extensions==2.2.9
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
fire==0.3.1
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
newscatcher==0.2.0
notebook==6.0.3
oauthlib==3.1.0
onnx==1.7.0
onnxconverter-common==1.7.0
opt-einsum==3.2.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.6.8
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
sqlparse==0.3.1
srsly==1.0.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tldextract==2.2.2
tokenizers==0.7.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
